+++ R/empty_space.R
@@ -1,0 +1,288 @@
+# THE FUNCTIONS ARE DESIGN TO FIND EMPTY SPACES 
+# I thin I found a solution to the problem 
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+# plot_void <- function(x,y, table.length = floor(sqrt(length(y))), plot=TRUE)
+# {
+#    xr <- range(x)
+#    yr <- range(y)
+#    xg <- cut(x, table.length) #seq(xr[1], xr[2], length.out=length+1))
+#    yg <- cut(y, table.length) #seq(yr[1], yr[2], length.out=length+1))
+#    df <- data.frame(x=xg, y=yg)
+#    tt <- table(df$x,df$y)
+# if (plot) image(tt)
+#   cat("% void", per = (sum(tt==0)/(table.length^2)), "\n")
+#   cat("table length =", table.length, "\n") 
+#   cat("prob0 =", exp(-mean(tt)), "\n") 
+#   cat("mean =", mean(tt), "\n") 
+#   invisible(tt)
+# }
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+# going through different lengths 
+# get the mean (meantt) of the table and find the probability for zero (prob0)
+# (perc) is the percentage of zero cells
+# 
+# find_void <- function(x,y, 
+#     table.length = c(5,10,15,20,25,30, 35,40,45,50, 60, 70,80,90,100),
+#             Plot = TRUE)
+# {
+# ################################################################################
+#   emply_spaces <- function(x,y, table.length = 10)
+#   {
+#      xr <- range(x)
+#     uxl <- (xr[2]-xr[1])/table.length
+#      yr <- range(y)
+#     uyl <- (yr[2]-yr[1])/table.length
+#      xg <- cut(x, table.length) #seq(xr[1], xr[2], length.out=length+1))
+#      yg <- cut(y, table.length) #seq(yr[1], yr[2], length.out=length+1))
+#      df <- data.frame(x=xg, y=yg)
+#      tt <- table(df$x,df$y)
+#  meantt <- mean(tt)
+#   prob0 <- exp(-meantt)
+#    perc <- sum(tt==0)/(table.length^2)
+#     pen <- abs(perc-(1-prob0))
+#   c(pen=pen, perc=perc, prob0=prob0)
+#   }
+# ################################################################################  
+#      lt <- length(table.length)
+#   prob0 <- rep(0, lt)
+#    pen <- rep(0, lt)
+#   perc <- rep(0, lt)
+#     ii <- 0
+# for(i in table.length)
+#   {
+#     ii <- ii+1
+#    CC <- emply_spaces(x,y, table.length=i)  
+#   pen[ii] <- CC[1]
+#  perc[ii] <- CC[2]
+# prob0[ii] <- CC[3]
+#   }
+# if (Plot) 
+#   {
+#   op<- par(mfrow=c(1,2))
+#   plot(perc~table.length)
+#   points(prob0~table.length, col="red")
+#   plot(pen~table.length)
+#   par(op)
+#   }  
+# cat("void =", 1-perc[which.min(pen)], "\n") 
+# cat("table length =", table.length[which.min(pen)], "\n") 
+# invisible(list(perc,pen, prob0))
+# } 
+################################################################################
+################################################################################
+################################################################################
+################################################################################   
+# data 
+# size<-c(100,200,300,400,500,600,700,800,900,1000,2000,3000,4000,5000,10000,20000,30000,40000,50000,100000, 500000,1000000)
+# log(length)~log(size)<-c(6,8,10,12,13,15,16,17,18,19,26,32,37,41,58,82,100,118,129,190,420,580)
+# plot(log(length)~log(size))
+# m1=lm(log(length)~log(size))
+# coef(m1)
+#  -0.5165698   0.4983922 
+#  
+################################################################################
+# THE FUNCTION VOID() i my latest idea of calculation  empty spaces ian x ans y direction 
+# For given x and y 
+# get the length of the variables , `n`,
+# for the number of variables calculates the length, of the table `table.length`,
+#  to use to calculate empty spaces
+# create a table of size `table.length*table.length`
+#  and calculates the number of zeros in the table.
+# the way  the length of the table is  calculated is like this 
+# i) for a given `n` use the function fn() to calculate the table length `table.length`
+# ii) the coefficients in fn() are obtain by fitting a linear model to log(length)
+# against log(size) of the data given above 
+# Th way the data were calculated is as follows 
+#  for given size create two runif(n) variables and calculare which values of 
+#  `table.length gives a close to 0.05 value       
+################################################################################
+################################################################################
+################################################################################
+################################################################################     
+# the function void() looks of % of empty spaces in the x and y direction
+################################################################################ 
+void <-  function(x,y,  plot=TRUE, print=TRUE)
+  {
+  n <-length(x)
+if (length(y)!=n) stop("the x and y should have the same length")  
+  fn<-function(n) exp(-0.5165698+0.4983922*log(n) )
+  table.length <- floor(fn(n))
+    xr <- range(x)
+    yr <- range(y)
+    xg <- cut(x, table.length) #seq(xr[1], xr[2], length.out=length+1))
+    yg <- cut(y, table.length) #seq(yr[1], yr[2], length.out=length+1))
+    df <- data.frame(x=xg, y=yg)
+    tt <- table(df$x,df$y)
+    if (plot) image(tt)
+    if (print)
+    {
+      cat("% void", (sum(tt==0)/(table.length^2)), "\n")
+      cat("table length =", table.length, "\n") 
+      cat("prob0 =", exp(-mean(tt)), "\n") 
+      cat("mean =", mean(tt), "\n") 
+    }
+    invisible((sum(tt==0)/(table.length^2)))
+}
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+#  empty spaces for a data.frame
+# it check whether they are factor and only show the
+#  empty spaces for continuous variables
+# The version uses foreach
+data_void <- function(data,   
+                    digits = 3,
+                      plot = TRUE,
+                  diag.off = TRUE,
+             lower.tri.off = FALSE,  
+                    method = c("square", "circle"),
+                fit.method = c("P-splines", "loess"),
+             outline.color = "gray",
+                    colors = c("blue", "white", "red"),
+              legend.title = "Void",
+                    title,
+                   ggtheme = ggplot2::theme_minimal(),
+                    tl.cex = 12,
+                    tl.col = "black", 
+                    tl.srt = 45,
+                       lab = TRUE, 
+                   lab_col = "black", 
+                  lab_size = 3,
+               circle.size = 20,
+                        ...) # c(1,15) maybe will do
+{
+################################################################################
+################################################################################
+#  require(foreach)
+################################################################################
+################################################################################
+  # local function 
+  meltit <- function(mat)
+  {
+    rna <- rownames(mat)
+    lrna <- length(rna)
+    value <- as.vector(mat)
+    Var1 <- gl(length(rna), 1, length = lrna*lrna, labels=rna)
+    Var2 <- gl(length(rna), lrna, length = lrna*lrna, labels=rna)
+    daf <-  na.omit(data.frame(Var1, Var2, value=value)) 
+    daf
+  }
+################################################################################ 
+################################################################################    
+# data.frame missing 
+     i <- j <- NULL
+if (missing(data) || NROW(data) <= 1) 
+    stop("nothing to do for this data frame")
+# data obs na's
+if (is(data, "list"))  stop("the data is list  the function needs a data.frame") 
+if (is(data, "table")) stop("the data is a table the function needs a data.frame")
+if (is(data, "matrix"))    data <- as.data.frame(data)
+if (is(data[1],"mts"))     data <- as.data.frame(data)
+if (is(data, "array")) stop("the data is an array the function needs a data.frame")    
+     dimD <- dim(data)
+if (any(is.na(data)))
+  {
+       l1 <- dim(data)[1]
+     data <- na.omit(data)
+       l2 <- dim(data)[1]
+    warning(cat(l1-l2, "observations were omitted from the data", "\n"))
+  }
+#  if is a list or table
+if (is.null(dimD)) stop("only one variable in the data") 
+if (dimD[1]<20)   stop(cat("the size of the data set is too small", "\n",
+                             "to detect empty spaces", "\n"))   
+       TT <- lapply(data, unique)
+     daTa <- subset(data, select=
+              ifelse(sapply(data,is.factor) | sapply(data,is.character) |
+                    sapply(data,is.integer)&lapply(TT, length)<10, FALSE, TRUE))
+      Dim <- dim(daTa)
+if (Dim[2]==0) stop("no variable is left after taking out the factors")             
+if (Dim[2]==1) stop("only one variable is left after taking out the factors")              
+ diffDim  <- dimD[2]-Dim[2]
+if (diffDim > 0)
+  {
+    warning(cat(diffDim, 'factors have been omited from the data', "\n"))
+  }
+   cnames <- names(daTa)
+  lcnames <- length(cnames)
+       CC <- matrix(0, ncol=lcnames, nrow=lcnames)
+  #get CC using foreach
+      CC <- foreach(i=1:lcnames, .combine='rbind') %dopar% 
+    {
+      xi <- if(is.null(dim(daTa[,i]))) daTa[,i] else daTa[,i][,1]
+      foreach(j=1:lcnames, .combine='c') %do% 
+        {
+          xj <- if(is.null(dim(daTa[,j]))) daTa[,j] else daTa[,j][,1] 
+          if (i<j) CC[i,j]  <- void(xi, xj, plot=FALSE, print=FALSE)
+          else CC[i,j] <- 0
+        }
+    }
+   CC <- CC+t(CC)  # to get the full matrix (rather than diagonal)
+#  CC1 <- cor(daTa)
+#  CC1 <- base::round(x = CC, digits = digits)            
+  rownames(CC) <- cnames
+  colnames(CC) <- cnames
+      diag(CC) <- 1
+            CC <- base::round(x = CC, digits = digits)   
+if (diag.off) diag(CC) <- NA
+if  (lower.tri.off) CC[lower.tri(CC)] <- NA
+if (plot==FALSE) return(CC)
+      method <- match.arg(method)
+  fit.method <- match.arg(fit.method)
+        corr <- meltit(CC)
+    lowerLim <- 25-floor((range(corr$value)[2]-range(corr$value)[1])*20)
+  colnames(corr) <- c("var_1", "var_2", "value")
+  txt.title <- if (missing(title))
+    paste("Empty spaces from data:", deparse(substitute(data)))
+  else title
+  corr$abs_corr <- abs(corr$value) * 10
+  p <- ggplot2::ggplot(data = corr,
+                       mapping = ggplot2::aes_string(x = "var_1", y = "var_2", 
+                                                     fill = "value"))
+  if (method == "square")
+  {
+    p <- p + ggplot2::geom_tile(color = outline.color)
+  }
+  else if (method == "circle") 
+  {
+    p <- p + geom_point(color = outline.color, shape = 21,
+                        ggplot2::aes_string(size = "abs_corr")) +
+      # scale_size(range = circle_scale_size_range) +
+      scale_size_area(max_size = circle.size) +
+      guides(size = "none")
+  }
+  label <- round(x = CC, digits = digits)
+  p <- p + ggplot2::scale_fill_gradient2(low = colors[1], high = colors[3],
+                                         mid = colors[2],  midpoint = 0.5, limit = c(0, 1),
+                                         space = "Lab",
+                                         name = legend.title)+ggplot2::ggtitle(txt.title)
+  if (class(ggtheme)[[1]] == "function") {
+    p <- p + ggtheme
+  }
+  else if (class(ggtheme)[[1]] == "theme") {
+    p <- p + ggtheme
+  }
+  p <- p + ggplot2::theme(axis.text.x = ggplot2::element_text(angle = tl.srt,
+                                                              vjust = 1, size = tl.cex, colour=tl.col, hjust = 1),
+                          axis.text.y = ggplot2::element_text(size = tl.cex, 
+                                                              colour=tl.col)) +
+    ggplot2::coord_fixed()
+  label <- round(x = corr[, "value"], digits = digits)
+  if (lab) {
+    p <- p + ggplot2::geom_text(
+      mapping = ggplot2::aes_string(x = "var_1", y = "var_2"),
+      label = label, color = lab_col, size = lab_size)
+  }
+  p
+}       
+################################################################################
+################################################################################
+################################################################################
+################################################################################